requirements.txt
gradio>=4.0.0
torch>=2.0.0
transformers>=4.44.0
accelerate
safetensors
pillow
numpy
